requirements.txt
bleach==4.0.0
Django==4.2.1
django-rq==2.5.1
lxml==4.8.0
python-ldap==3.4.0
pyvcloud==23.0.3
pyvmomi==7.0.3
PyYAML==6.0
redis==3.5.3
rq==1.11.1
urllib3==1.26.8
django-redis-cache==3.0.1
docutils==0.19
djangorestframework==3.14.0
django-debug-toolbar==4.1.0
django-jet-reboot==1.3.3
django-crontab==0.7.1
whitenoise==6.6.0
gunicorn==21.2.0